+++ comfy_extras/nodes_wan.py
@@ -1456,161 +1456,6 @@ class WanInfiniteTalkToVideo(io.ComfyNode):
         return io.NodeOutput(model_patched, positive, negative, out_latent, trim_image)
 
 
-def _extract_mask_to_28ch(rgb_video):
-    """Colored RGB mask (T, H, W, 3) in [0, 1] -> SCAIL-2 28-channel binary latent
-    (1, T_lat, 28, H_lat, W_lat). 7 per-color binary channels (white/r/g/b/y/m/c)
-    threshold-extracted at 225/255, 8x spatial downsample, 4-frame temporal stacking."""
-    T, H, W, _ = rgb_video.shape
-    _ON_THRESH = 225.0 / 255.0
-    mask = rgb_video.movedim(-1, 1).float()
-    R = (mask[:, 0:1] > _ON_THRESH).float()
-    G = (mask[:, 1:2] > _ON_THRESH).float()
-    B = (mask[:, 2:3] > _ON_THRESH).float()
-    nR, nG, nB = 1 - R, 1 - G, 1 - B
-    binary_7ch = torch.cat([
-        R * G * B,    # white
-        R * nG * nB,  # red
-        nR * G * nB,  # green
-        nR * nG * B,  # blue
-        R * G * nB,   # yellow
-        R * nG * B,   # magenta
-        nR * G * B,   # cyan
-    ], dim=1)
-    H_lat, W_lat = H, W
-    for _ in range(3):
-        H_lat = (H_lat + 1) // 2
-        W_lat = (W_lat + 1) // 2
-    binary_7ch = torch.nn.functional.interpolate(binary_7ch, size=(H_lat, W_lat), mode='area')
-    T_latent = (T - 1) // 4 + 1
-    padded = torch.cat([binary_7ch[:1].repeat(4, 1, 1, 1), binary_7ch[1:]], dim=0)
-    out = padded.view(T_latent, 28, H_lat, W_lat)
-    return out.unsqueeze(0)
-
-
-class WanSCAILToVideo(io.ComfyNode):
-    @classmethod
-    def define_schema(cls):
-        return io.Schema(
-            node_id="WanSCAILToVideo",
-            category="model/conditioning/video_models",
-            inputs=[
-                io.Conditioning.Input("positive"),
-                io.Conditioning.Input("negative"),
-                io.Vae.Input("vae"),
-                io.Int.Input("width", default=512, min=32, max=nodes.MAX_RESOLUTION, step=32),
-                io.Int.Input("height", default=896, min=32, max=nodes.MAX_RESOLUTION, step=32),
-                io.Int.Input("length", default=81, min=1, max=nodes.MAX_RESOLUTION, step=4),
-                io.Int.Input("batch_size", default=1, min=1, max=4096),
-                io.Image.Input("pose_video", optional=True, tooltip="Video used for pose conditioning. Will be downscaled to half the resolution of the main video."),
-                io.Image.Input("driving_mask_video", optional=True, tooltip="SCAIL-2 only. Colored per-identity SAM3 mask video at the same resolution as pose_video. Mode is auto-detected from bg color: black bg = Animation, white bg = Replacement."),
-                io.Float.Input("pose_strength", default=1.0, min=0.0, max=10.0, step=0.01, tooltip="Strength of the pose latent."),
-                io.Float.Input("pose_start", default=0.0, min=0.0, max=1.0, step=0.01, tooltip="Start step to use pose conditioning."),
-                io.Float.Input("pose_end", default=1.0, min=0.0, max=1.0, step=0.01, tooltip="End step to use pose conditioning."),
-                io.Image.Input("reference_image", optional=True),
-                io.Image.Input("ref_mask_image", optional=True, tooltip="SCAIL-2 only. Single-frame colored ref mask at the reference image's full resolution."),
-                io.ClipVisionOutput.Input("clip_vision_output", optional=True),
-                io.Int.Input("video_frame_offset", default=0, min=0, max=nodes.MAX_RESOLUTION, step=1, tooltip="Cumulative output frame this chunk begins at. Wire from the previous chunk's video_frame_offset output."),
-                io.Int.Input("previous_frame_count", default=5, min=1, max=nodes.MAX_RESOLUTION, step=4, tooltip="Tail frames of previous_frames to anchor. SCAIL-2 trained at 5 (81-frame chunks, 76-frame step)."),
-                io.Image.Input("previous_frames", optional=True, tooltip="SCAIL-2 only. Full decoded output of the previous chunk. Only the last previous_frame_count are used as the inpainting anchor."),
-            ],
-            outputs=[
-                io.Conditioning.Output(display_name="positive"),
-                io.Conditioning.Output(display_name="negative"),
-                io.Latent.Output(display_name="latent", tooltip="Empty latent of the generation size."),
-                io.Int.Output(display_name="video_frame_offset", tooltip="Adjusted offset + length. Wire into the next chunk."),
-            ],
-            is_experimental=True,
-        )
-
-    @classmethod
-    def execute(cls, positive, negative, vae, width, height, length, batch_size,
-                pose_strength, pose_start, pose_end,
-                video_frame_offset, previous_frame_count,
-                reference_image=None, clip_vision_output=None, pose_video=None,
-                driving_mask_video=None, ref_mask_image=None, previous_frames=None) -> io.NodeOutput:
-        latent = torch.zeros([batch_size, 16, ((length - 1) // 4) + 1, height // 8, width // 8], device=comfy.model_management.intermediate_device())
-        noise_mask = None
-
-        # Auto-detect mode from driving_mask_video bg color. White bg => Replacement, else Animation.
-        replacement_mode = driving_mask_video is not None and driving_mask_video[0, ..., :3].mean().item() > 0.5
-        ref_mask_flag = torch.tensor([not replacement_mode], dtype=torch.bool, device=latent.device)
-        positive = node_helpers.conditioning_set_values(positive, {"ref_mask_flag": ref_mask_flag})
-        negative = node_helpers.conditioning_set_values(negative, {"ref_mask_flag": ref_mask_flag})
-
-        prev_trimmed = None
-        if previous_frames is not None and previous_frames.shape[0] > 0:
-            prev_trimmed = previous_frames[-previous_frame_count:]
-            video_frame_offset -= prev_trimmed.shape[0]
-            video_frame_offset = max(0, video_frame_offset)
-
-        ref_latent = None
-        if reference_image is not None:
-            reference_image = comfy.utils.common_upscale(reference_image[:1].movedim(-1, 1), width, height, "bilinear", "center").movedim(1, -1)
-            # Replacement Mode: composite ref on black bg using ref_mask_image as alpha matte
-            # (matches the pre-composited examples that ship with SCAIL-2). Pixels where the
-            # mask is non-black (max channel > 0.1) are kept; bg pixels go to black.
-            if replacement_mode and ref_mask_image is not None:
-                rm = comfy.utils.common_upscale(ref_mask_image[:1].movedim(-1, 1), width, height, "nearest-exact", "center").movedim(1, -1)
-                is_char = (rm[..., :3].max(dim=-1, keepdim=True).values > 0.1).to(reference_image.dtype)
-                reference_image = reference_image * is_char
-            ref_latent = vae.encode(reference_image[:, :, :, :3])
-
-        if ref_latent is not None:
-            positive = node_helpers.conditioning_set_values(positive, {"reference_latents": [ref_latent]}, append=True)
-            negative = node_helpers.conditioning_set_values(negative, {"reference_latents": [ref_latent]}, append=True)
-
-        if clip_vision_output is not None:
-            positive = node_helpers.conditioning_set_values(positive, {"clip_vision_output": clip_vision_output})
-            negative = node_helpers.conditioning_set_values(negative, {"clip_vision_output": clip_vision_output})
-
-        if pose_video is not None:
-            if pose_video.shape[0] <= video_frame_offset:
-                pose_video = None
-            else:
-                pose_video = pose_video[video_frame_offset:]
-        if driving_mask_video is not None:
-            if driving_mask_video.shape[0] <= video_frame_offset:
-                driving_mask_video = None
-            else:
-                driving_mask_video = driving_mask_video[video_frame_offset:]
-
-        if pose_video is not None:
-            pose_video = comfy.utils.common_upscale(pose_video[:length].movedim(-1, 1), width // 2, height // 2, "area", "center").movedim(1, -1)
-            pose_video_latent = vae.encode(pose_video[:, :, :, :3]) * pose_strength
-            positive = node_helpers.conditioning_set_values_with_timestep_range(positive, {"pose_video_latent": pose_video_latent}, pose_start, pose_end)
-            negative = node_helpers.conditioning_set_values_with_timestep_range(negative, {"pose_video_latent": pose_video_latent}, pose_start, pose_end)
-
-        if driving_mask_video is not None:
-            mask_video_hw = comfy.utils.common_upscale(driving_mask_video[:length].movedim(-1, 1), width // 2, height // 2, "area", "center").movedim(1, -1)
-            sam_28ch = _extract_mask_to_28ch(mask_video_hw)
-            positive = node_helpers.conditioning_set_values(positive, {"sam_28ch": sam_28ch})
-            negative = node_helpers.conditioning_set_values(negative, {"sam_28ch": sam_28ch})
-
-        if ref_mask_image is not None:
-            ref_mask_hw = comfy.utils.common_upscale(ref_mask_image[:1].movedim(-1, 1), width, height, "bilinear", "center").movedim(1, -1)
-            ref_sam_1f = _extract_mask_to_28ch(ref_mask_hw)
-            T_lat = ((length - 1) // 4) + 1
-            zeros = torch.zeros((1, T_lat, 28, ref_sam_1f.shape[-2], ref_sam_1f.shape[-1]),
-                                device=ref_sam_1f.device, dtype=ref_sam_1f.dtype)
-            ref_sam_28ch = torch.cat([ref_sam_1f, zeros], dim=1)
-            positive = node_helpers.conditioning_set_values(positive, {"ref_sam_28ch": ref_sam_28ch})
-            negative = node_helpers.conditioning_set_values(negative, {"ref_sam_28ch": ref_sam_28ch})
-
-        if prev_trimmed is not None:
-            pf = comfy.utils.common_upscale(prev_trimmed.movedim(-1, 1), width, height, "bilinear", "center").movedim(1, -1)
-            prev_latent = vae.encode(pf[:, :, :, :3])
-            T_p_lat = min(prev_latent.shape[2], latent.shape[2])
-            latent[:, :, :T_p_lat] = prev_latent[:, :, :T_p_lat].to(latent.dtype)
-            noise_mask = torch.ones((1, 1, latent.shape[2], latent.shape[-2], latent.shape[-1]),
-                                    device=latent.device, dtype=latent.dtype)
-            noise_mask[:, :, :T_p_lat] = 0.0
-
-        out_latent = {"samples": latent}
-        if noise_mask is not None:
-            out_latent["noise_mask"] = noise_mask
-        return io.NodeOutput(positive, negative, out_latent, video_frame_offset + length)
-
-
 class WanExtension(ComfyExtension):
     @override
     async def get_node_list(self) -> list[type[io.ComfyNode]]:
@@ -1631,7 +1476,6 @@ class WanExtension(ComfyExtension):
             WanAnimateToVideo,
             Wan22ImageToVideoLatent,
             WanInfiniteTalkToVideo,
-            WanSCAILToVideo,
         ]
 
 async def comfy_entrypoint() -> WanExtension:
